feat(kmp): looper tanda 1 — las 40 que ya tenian C API, y un int que angostaba
Primera de cuatro tandas del looper, la ultima categoria de WA-2.5/2.6.
42 entry points del JNI pasan a delegar (delegacion 148 -> 190/278, looper
1/79 -> 43/79). Cero funciones nuevas de C API: estas 40 ya existian y el JNI
las transcribia a mano igual.

Los 40 pares se diffearon lado a lado antes de migrar: transcripciones fieles,
cero divergencia semantica. prepareTrack cruza dos tablas de error
(JniError vs WmaResult) y son identicas entrada por entrada, asi que la
delegacion preserva el valor que lee NativeErrorCode del lado Kotlin.

BUG QUE LA MIGRACION MECANICA IBA A INTRODUCIR, visto al leer el diff:
wma_looper_set_track_loop_region declaraba `int start_frame, int end_frame`,
pero la cadena es 64-bit de punta a punta — Kotlin Long, jlong en el JNI, y
AudioLooper::setTrackLoopRegion toma int64_t y satura a int32 el mismo, con un
comentario diciendo que el ancho es deliberado porque TrackBuffer guarda
frames como int32. La C API era el unico eslabon angosto: la saturacion nunca
podia correr porque el valor llegaba ya truncado. Pasar los jlong a la firma
vieja habria hecho que el JNI perdiera el ensanchado. Ensanchada a int64_t.

nativeLooperGetTrackWaveform ahora clampea al largo real del array de Java,
como su hermano nativeGetWaveformSamples, que lo hace y explica por que.
No era explotable: su unico caller aloca FloatArray(numBins) y pasa el mismo
numBins. Deja de depender de que los dos numeros se mantengan en sync solos.

test_c_api_looper.cpp, 12 tests (679 en total). El foco NO es AudioLooper
—looper/tests/ ya lo cubre mejor— sino la frontera de la C API: el barrido de
null-handle pincha los 38 defaults que se movieron del JNI (los de 1.0f, que
son los que un default zero-initialized rompe sin que se note), el int64 del
loop region, y un ida-y-vuelta AUDIBLE que graba salida real del motor y la
reproduce, para probar que la C API llega al mismo AudioLooper que renderiza
el callback — el modo de falla del InputNode duplicado.

Tres intentos de test que fallaron por razon equivocada y quedaron anotados:
el loop region no existe en un track sin largo grabado (prepareTrack reserva
capacidad, mLengthFrames queda en 0); no se puede afirmar "el sinte se callo"
mientras el loop suena, asi que va pause/resume primero; y el mute esta
smootheado (~0.28 de ganancia despues del primer bloque de 256), asi que medir
al toque lee la rampa y no el mute. 3 mutantes, 3 detectados.

Gate local completo en verde: portabilidad, 679 tests C++, ambos slices de
iOS con link check, 87 tests de simulador, 50 JVM, assembleDebug, XCFramework
y compileIosMainKotlinMetadata.

+++ audio/src/main/cpp/api/watermelon_audio.cpp
@@ -1480,7 +1480,7 @@ float wma_looper_get_master_volume(const WmaEngine* engine) {
 }
 
 void wma_looper_set_track_loop_region(WmaEngine* engine, int track_index,
-                                       int start_frame, int end_frame) {
+                                       int64_t start_frame, int64_t end_frame) {
     WMA_CHECK_VOID(engine);
     engine->engine->getAudioLooper().setTrackLoopRegion(track_index, start_frame, end_frame);
 }
